refactor: 🔨 electron-folder
