+++ electron/main.ts
@@ -1,45 +1,39 @@
-import { app, BrowserWindow, ipcMain, nativeImage, Menu } from 'electron'
+import { app, BrowserWindow, ipcMain, nativeImage, Menu, MenuItemConstructorOptions } from 'electron'
 import path from 'node:path'
 
 process.env.DIST = path.join(__dirname, '../dist')
 process.env.PUBLIC = app.isPackaged ? process.env.DIST : path.join(process.env.DIST, '../public')
 
-
 let win: BrowserWindow | null
 
-const VITE_DEV_SERVER_URL = process.env['VITE_DEV_SERVER_URL']
+const VITE_DEV_SERVER_URL = process.env.VITE_DEV_SERVER_URL
 
-
-function createWindow() {
+function createWindow () {
   win = new BrowserWindow({
     icon: path.join(process.env.PUBLIC, 'jsrunner.png'),
     frame: false,
     webPreferences: {
       preload: path.join(__dirname, 'preload.js'),
       devTools: true,
-      nodeIntegration: true,
+      nodeIntegration: true
 
-    },
+    }
 
   })
   ipcMain.on('close-me', () => {
-
-    app.quit();
-  });
+    app.quit()
+  })
   ipcMain.on('maximize', () => {
-
-    win?.maximize();
-  });
+    win?.maximize()
+  })
   ipcMain.on('unmaximize', () => {
-
-    win?.unmaximize();
-  });
-
+    win?.unmaximize()
+  })
 
   // TODO: CONNECT THE MENU WITH NPM WORKSPACE
 
   ipcMain.on('show-context-menu', (event) => {
-    const template = [
+    const template: MenuItemConstructorOptions[] = [
       {
         label: 'Working on npm workspace module',
         click: () => { event.sender.send('context-menu-command', 'menu-item-1') }
@@ -48,12 +42,12 @@ function createWindow() {
       { label: 'coming soon', type: 'checkbox', checked: false }
     ]
     const menu = Menu.buildFromTemplate(template)
-    menu.popup({ window: BrowserWindow.fromWebContents(event.sender) })
+    menu.popup({ window: BrowserWindow.fromWebContents(event.sender) ?? undefined })
   })
 
   // Test active push message to Renderer-process.
   win.webContents.on('did-finish-load', () => {
-    win?.webContents.send('main-process-message', (new Date).toLocaleString())
+    win?.webContents.send('main-process-message', (new Date()).toLocaleString())
   })
 
   if (VITE_DEV_SERVER_URL) {
